+++ Axi4/testbench_interrupt/TbAxi4Memory.vhd
@@ -134,7 +134,7 @@ begin
 
   -- create Clock
   ------------------------------------------------------------
-  Osvvm.TbUtilPkg.CreateClock (
+  Osvvm.ClockResetPkg.CreateClock (
   ------------------------------------------------------------
     Clk        => Clk,
     Period     => Tperiod_Clk
@@ -142,7 +142,7 @@ begin
 
   -- create nReset
   ------------------------------------------------------------
-  Osvvm.TbUtilPkg.CreateReset (
+  Osvvm.ClockResetPkg.CreateReset (
   ------------------------------------------------------------
     Reset       => nReset,
     ResetActive => '0',