+++ Axi4/testbench_xilinx/TbAxi4.vhd
@@ -132,13 +132,13 @@ architecture TestHarness of TbAxi4 is
 begin
 
   -- create Clock
-  Osvvm.TbUtilPkg.CreateClock (
+  Osvvm.ClockResetPkg.CreateClock (
     Clk        => Clk,
     Period     => Tperiod_Clk
   )  ;
 
   -- create nReset
-  Osvvm.TbUtilPkg.CreateReset (
+  Osvvm.ClockResetPkg.CreateReset (
     Reset       => nReset,
     ResetActive => '0',
     Clk         => Clk,