+++ AxiStream/TestCases/TbStream_WaitForGet2.vhd
@@ -1,5 +1,5 @@
 --
---  File Name:         TbStream_MultipleDriversReceiver1.vhd
+--  File Name:         TbStream_WaitForGet2.vhd
 --  Design Unit Name:  Architecture of TestCtrl
 --  Revision:          OSVVM MODELS STANDARD VERSION
 --
@@ -9,7 +9,8 @@
 --
 --
 --  Description:
---      Validates Multiple Driver detection works
+--      Validates RECEIVE_READY_WAIT_FOR_GET with 
+--      word and burst receive transactions
 --
 --
 --  Developed by:
@@ -19,29 +20,29 @@
 --
 --  Revision History:
 --    Date      Version    Description
---    02/2021   2021.02    Initial revision
+--    01/2022   2022.01    Initial revision
 --
 --
 --  This file is part of OSVVM.
---  
---  Copyright (c) 2021 by SynthWorks Design Inc.  
---  
+--
+--  Copyright (c) 2022 by SynthWorks Design Inc.
+--
 --  Licensed under the Apache License, Version 2.0 (the "License");
 --  you may not use this file except in compliance with the License.
 --  You may obtain a copy of the License at
---  
+--
 --      https://www.apache.org/licenses/LICENSE-2.0
---  
+--
 --  Unless required by applicable law or agreed to in writing, software
 --  distributed under the License is distributed on an "AS IS" BASIS,
 --  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 --  See the License for the specific language governing permissions and
 --  limitations under the License.
---  
-architecture MultipleDriversReceiver1 of TestCtrl is
+--
+architecture WaitForGet2 of TestCtrl is
 
-  signal   TestDone : integer_barrier := 1 ;
-   
+  signal   Sync     : integer_barrier := 1 ;
+
 begin
 
   ------------------------------------------------------------
@@ -51,73 +52,88 @@ begin
   ControlProc : process
   begin
     -- Initialization of test
-    SetTestName("TbStream_MultipleDriversReceiver1") ;
+    SetTestName("TbStream_WaitForGet2") ;
     SetLogEnable(PASSED, TRUE) ;    -- Enable PASSED logs
     SetLogEnable(INFO, TRUE) ;    -- Enable INFO logs
-    SetAlertStopCount(FAILURE, 2) ;    -- Allow 2 FAILURE Alerts
 
-    -- Wait for testbench initialization 
+    -- Wait for simulation elaboration/initialization
     wait for 0 ns ;  wait for 0 ns ;
     TranscriptOpen ;
-    SetTranscriptMirror(TRUE) ; 
+    SetTranscriptMirror(TRUE) ;
 
     -- Wait for Design Reset
-    wait until nReset = '1' ;  
+    wait until nReset = '1' ;
     ClearAlerts ;
 
     -- Wait for test to finish
     WaitForBarrier(TestDone, 35 ms) ;
     AlertIf(now >= 35 ms, "Test finished due to timeout") ;
---    AlertIf(GetAffirmCount < 1, "Test is not Self-Checking");
-    
-    AffirmIf(GetAlertCount = AlertCountType'(FAILURE => 1, ERROR => 0, WARNING => 0), "Expecting: (FAILURE => 1, ERROR => 0, WARNING => 0)") ;
-    TranscriptClose ; 
-    AffirmIfTranscriptsMatch(OSVVM_VALIDATED_RESULTS_DIR) ;   
-   
-   
-    
-    EndOfTestReports(ExternalErrors => (FAILURE => -1, ERROR => 0, WARNING => 0)) ; 
-    std.env.stop ;
-    wait ; 
-  end process ControlProc ; 
+    AlertIf(GetAffirmCount < 1, "Test is not Self-Checking");
 
-  
+    TranscriptClose ;
+    -- AffirmIfTranscriptsMatch(OSVVM_VALIDATED_RESULTS_DIR) ;   
+   
+    EndOfTestReports(ExternalErrors => (0, 0, 0)) ;
+    std.env.stop ;
+    wait ;
+  end process ControlProc ;
+
+
   ------------------------------------------------------------
-  -- TransmitterProc
-  --   Generate transactions for Transmitter
+  -- AxiTransmitterProc
+  --   Generate transactions for AxiTransmitter
   ------------------------------------------------------------
   TransmitterProc : process
+    variable CoverID : CoverageIdType ;
   begin
-    wait until nReset = '1' ;  
-    WaitForClock(StreamTxRec, 2) ; 
-    WaitForClock(StreamTxRec, 2) ; 
-    WaitForClock(StreamRxRec, 2) ; 
 
+    WaitForClock(StreamTxRec, 2) ;
+
+-- Send and Get
+    BlankLine(2) ; 
+    log("Send and Get: Transmit 4 words") ;
+    for I in 0 to 3 loop
+      Send( StreamTxRec, X"0000_1000" + I ) ;
+    end loop ;
+
+    -- Wait for outputs to propagate and signal TestDone
+    WaitForClock(StreamTxRec, 2) ;
     WaitForBarrier(TestDone) ;
     wait ;
   end process TransmitterProc ;
 
 
   ------------------------------------------------------------
-  -- ReceiverProc
-  --   Generate transactions for Receiver
+  -- AxiReceiverProc
+  --   Generate transactions for AxiReceiver
   ------------------------------------------------------------
   ReceiverProc : process
+    variable RxData : std_logic_vector(DATA_WIDTH-1 downto 0) ;
+    variable NumWords : integer ;
+    variable GetStartTime  : time ;
   begin
-    wait until nReset = '1' ;  
-    WaitForClock(StreamRxRec, 2) ; 
-    WaitForClock(StreamRxRec, 3) ; 
+    WaitForClock(StreamRxRec, 6) ;
+    SetAxiStreamOptions(StreamRxRec, RECEIVE_READY_WAIT_FOR_GET, TRUE) ;
     
+    WaitForBarrier(VcInit) ; 
+    
+    for I in 0 to 3 loop
+      Check(StreamRxRec, X"0000_1000" + I) ;
+      WaitForClock(StreamRxRec, 2) ;
+    end loop ;
+
+    -- Wait for outputs to propagate and signal TestDone
+    WaitForClock(StreamRxRec, 2) ;
     WaitForBarrier(TestDone) ;
     wait ;
   end process ReceiverProc ;
 
-end MultipleDriversReceiver1 ;
+end WaitForGet2 ;
 
-Configuration TbStream_MultipleDriversReceiver1 of TbStream is
+Configuration TbStream_WaitForGet2 of TbStream is
   for TestHarness
     for TestCtrl_1 : TestCtrl
-      use entity work.TestCtrl(MultipleDriversReceiver1) ; 
-    end for ; 
-  end for ; 
-end TbStream_MultipleDriversReceiver1 ; 
+      use entity work.TestCtrl(WaitForGet2) ;
+    end for ;
+  end for ;
+end TbStream_WaitForGet2 ;